plugins/chatGPT/pyproject.toml
[tool.poetry]
name = "nonebot-plugin-multi-chatgpt"
version = "2.0.3"
description = ""
authors = ["chrisyy <1017975501@qq.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_multi_chatgpt"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc2"
nonebot-adapter-onebot = "^2.1.5"
nonebot-plugin-apscheduler = "^0.2.0"
nonebot-plugin-htmlrender="^0.2.0.1"
revchatgpt = "^0.0.43"
pyyaml = "^5.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
